docs:刪除及重新命名檔案
調整項目:
1. 移除無用檔案
2. Instant Chatting重新命名
